tests/test_polymorphic_code_gen.py
import unittest
from polymorphic_code_gen import generate_polymorphic_code

class TestPolymorphicCodeGeneration(unittest.TestCase):
    def setUp(self):
        # Setup can include configuration settings or other necessary variables
        self.api_key = 'your_api_key_here'
        self.original_code = 'print("Hello, World!")'
        self.variation_count = 5  # Number of variations to generate for testing

    def test_generate_polymorphic_code(self):
        # Test if the polymorphic code generator is creating different variations
        variations = set()
        for _ in range(self.variation_count):
            code_variation = generate_polymorphic_code(self.original_code, self.api_key)
            self.assertNotIn(code_variation, variations, "Generated code should be unique")
            variations.add(code_variation)

    def test_generated_code_executes(self):
        # Test if the generated polymorphic code executes without errors
        for _ in range(self.variation_count):
            code_variation = generate_polymorphic_code(self.original_code, self.api_key)
            exec_globals = {}
            try:
                exec(code_variation, exec_globals)
            except Exception as e:
                self.fail(f"Execution of polymorphic code failed with error: {e}")

    def test_generated_code_preserves_functionality(self):
        # Test if the generated polymorphic code preserves the original functionality
        for _ in range(self.variation_count):
            code_variation = generate_polymorphic_code(self.original_code, self.api_key)
            exec_globals = {'__builtins__': __builtins__}
            exec(code_variation, exec_globals)
            self.assertIn('Hello, World!', exec_globals.get('output', ''),
                          "Generated code should preserve the original functionality")

if __name__ == '__main__':
    unittest.main()